+++ CHANGELOG.md
@@ -1,10 +1,16 @@
 # Changelog
 
-## 1.2.0 (2023-06-13)
+## 1.2.0 (2023-06-29)
 
-Add support for installing the CLI via NPM - [42](https://github.com/bugsnag/bugsnag-cli/pull/41)
+Add support for installing the CLI via NPM - [40](https://github.com/bugsnag/bugsnag-cli/pull/40)
 
-## 1.1.1 (2023-05-25)
+Move global `appVersion`, `appVersionCode` and `appBundleVersion` flags to sub commands for `dart` and `create-build` - [41](https://github.com/bugsnag/bugsnag-cli/pull/41)
+
+Correct `buildUUID` name in server requests for Android Proguard - [41](https://github.com/bugsnag/bugsnag-cli/pull/41)
+
+Get values from Android AAB manifest via resource ID - [41](https://github.com/bugsnag/bugsnag-cli/pull/41)
+
+## 1.1.1 (2023-05-25)   
 
 Fix how we check for the AndroidManifest.xml file for Android AAB - [37](https://github.com/bugsnag/bugsnag-cli/pull/37)